# docs/v1/swagger.yaml
basePath: /v1
definitions:
...
  request.CustomToolCreate:
    properties:
      apiAuth:
        allOf:
        - $ref: '#/definitions/request.CustomToolApiAuthWebRequest'
        description: api身份认证
      description:
        description: 描述
        type: string
      name:
        description: 名称
        type: string
      privacyPolicy:
        description: 隐私政策
        type: string
      schema:
        description: schema
        type: string
    required:
    - apiAuth
    - description
    - name
    - schema
    type: object
  request.CustomToolIDReq:
    properties:
      customToolId:
        description: 自定义工具id
        type: string
    required:
    - customToolId
    type: object
  request.CustomToolSchemaReq:
    properties:
      schema:
        description: schema
        type: string
    required:
    - schema
    type: object
  request.CustomToolUpdateReq:
    properties:
      apiAuth:
        allOf:
        - $ref: '#/definitions/request.CustomToolApiAuthWebRequest'
        description: api身份认证
      customToolId:
        description: 自定义工具ID
        type: string
      description:
        description: 描述
        type: string
      name:
        description: 名称
        type: string
      privacyPolicy:
        description: 隐私政策
        type: string
      schema:
        description: schema
        type: string
    required:
    - apiAuth
    - customToolId
    - description
    - name
    - schema
    type: object
...
  request.MCPServerCreateReq:
    properties:
      avatar:
        allOf:
        - $ref: '#/definitions/request.Avatar'
        description: 图标
      desc:
        description: 描述
        type: string
      name:
        description: 名称
        type: string
    required:
    - desc
    - name
    type: object
  request.MCPServerIDReq:
    properties:
      mcpServerId:
        type: string
    required:
    - mcpServerId
    type: object
  request.MCPServerOpenAPIToolCreate:
    properties:
      apiAuth:
        allOf:
        - $ref: '#/definitions/request.CustomToolApiAuthWebRequest'
        description: api身份认证
      mcpServerId:
        description: mcp server Id
        type: string
      methodNames:
        description: API名称列表
        items:
          type: string
        type: array
      name:
        description: 名称
        type: string
      privacyPolicy:
        description: 隐私政策
        type: string
      schema:
        description: schema
        type: string
    required:
    - apiAuth
    - mcpServerId
    - methodNames
    - name
    - schema
    type: object
  request.MCPServerToolCreateReq:
    properties:
      id:
        description: 应用或工具id
        type: string
      mcpServerId:
        description: mcp server Id
        type: string
      methodName:
        description: 显示名称
        type: string
      type:
        description: mcp server tool类型
        type: string
    required:
    - id
    - mcpServerId
    - methodName
    - type
    type: object
  request.MCPServerToolIDReq:
    properties:
      mcpServerToolId:
        description: mcp server tool id
        type: string
    required:
    - mcpServerToolId
    type: object
  request.MCPServerToolUpdateReq:
    properties:
      desc:
        description: 描述
        type: string
      mcpServerToolId:
        description: mcp server tool id
        type: string
      methodName:
        description: 显示名称
        type: string
    required:
    - desc
    - mcpServerToolId
    - methodName
    type: object
  request.MCPServerUpdateReq:
    properties:
      avatar:
        allOf:
        - $ref: '#/definitions/request.Avatar'
        description: 图标
      desc:
        description: 描述
        type: string
      mcpServerId:
        description: mcp server Id
        type: string
      name:
        description: 名称
        type: string
    required:
    - desc
    - mcpServerId
    - name
    type: object
...
  request.ToolSquareAPIKeyReq:
    properties:
      apiKey:
        description: apiKey
        type: string
      toolSquareId:
        description: 广场toolId
        type: string
    required:
    - toolSquareId
    type: object
...
  response.CustomToolDetail:
    properties:
      apiAuth:
        allOf:
        - $ref: '#/definitions/request.CustomToolApiAuthWebRequest'
        description: apiAuth
      apiList:
        description: action列表
        items:
          $ref: '#/definitions/response.CustomToolActionInfo'
        type: array
      customToolId:
        description: 自定义工具id
        type: string
      description:
        description: 描述
        type: string
      name:
        description: 名称
        type: string
      privacyPolicy:
        description: 隐私政策
        type: string
      schema:
        description: schema
        type: string
      toolSquareId:
        description: 广场mcpId(非空表示来源于广场)
        type: string
    type: object
  response.CustomToolInfo:
    properties:
      customToolId:
        description: 自定义工具id
        type: string
      description:
        description: 描述
        type: string
      name:
        description: 名称
        type: string
    type: object
...
  response.ListResult:
    properties:
      list: {}
      total:
        type: integer
    type: object
...
  response.MCPServerCustomToolSelect:
    properties:
      customToolId:
        description: 自定义工具id
        type: string
      description:
        description: 描述
        type: string
      methods:
        description: 方法
        items:
          $ref: '#/definitions/response.MCPServerCustomToolApi'
        type: array
      name:
        description: 名称
        type: string
      uniqueId:
        description: 统一的id
        type: string
    type: object
  response.MCPServerDetail:
    properties:
      avatar:
        allOf:
        - $ref: '#/definitions/request.Avatar'
        description: 图标
      desc:
        description: 描述
        type: string
      mcpServerId:
        description: mcpServerId
        type: string
      name:
        description: 名称
        type: string
      sseExample:
        description: sse连接示例
        type: string
      sseUrl:
        description: sse url
        type: string
      streamableExample:
        description: streamable http 连接示例
        type: string
      streamableUrl:
        description: streamable http url
        type: string
      tools:
        description: 绑定工具列表
        items:
          $ref: '#/definitions/response.MCPServerToolInfo'
        type: array
    type: object
  response.MCPServerInfo:
    properties:
      avatar:
        allOf:
        - $ref: '#/definitions/request.Avatar'
        description: 图标
      desc:
        description: 描述
        type: string
      mcpServerId:
        description: mcpServerId
        type: string
      name:
        description: 名称
        type: string
      toolNum:
        description: 绑定工具数量
        type: integer
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  response.ToolActionDetail:
    properties:
      action:
        allOf:
        - $ref: '#/definitions/protocol.Tool'
        description: action列表
      apiKey:
        description: apiKey
        type: string
      needApiKeyInput:
        description: 是否需要apiKey输入
        type: boolean
    type: object
  response.ToolActionList:
    properties:
      actions:
        description: action列表
        items:
          $ref: '#/definitions/protocol.Tool'
        type: array
    type: object
...
info:
  contact: {}
  description: "## HTTP Header\r\n| Header        | 说明      |\r\n| ------------- |
    --------- |\r\n| Authorization | JWT token |\r\n| X-Language    | 语言Code  |\r\n|
    X-Org-Id      | 组织ID    |\r\n| X-Client-Id   | 客户端标识|\r\n\r\n## HTTP Status\r\n|
    HTTP Status             | 说明                   |\r\n| -----------------------
    | ---------------------- |\r\n| 200, StatusOK           | 请求返回成功           |\r\n|
    400, StatusBadRequest   | 请求返回失败，用于业务 |\r\n| 401, StatusUnauthorized | JWT认证失败
    \           |\r\n| 403, StatusForbidden    | 没有权限               |\r\n\r\n## 权限-菜单对应表\r\n|
    一级权限        | 二级权限  | 三级权限 | 一级菜单 | 二级菜单 | 三级菜单 |\r\n|-------------|-------|------|------|------|------|\r\n|
    guest       |       |      | 【访客】 |      |      |\r\n| common      |       |      |
    【通用】 |      |      |\r\n| permission  |       |      | 权限管理 |      |      |\r\n|
    permission  | user  |      | 权限管理 | 用户管理 |      |\r\n| permission  | org   |      |
    权限管理 | 组织管理 |      |\r\n| permission  | role  |      | 权限管理 | 角色管理 |      |\r\n\r\n##
    `/v1/user/permission`返回用例\r\n```json\r\n{\r\n  \"code\": 0,\r\n  \"data\": {\r\n
    \   \"orgPermission\": {\r\n      \"org\": {\"id\": \"test-org-id\", \"name\":
    \"test-org-name\"},\r\n      \"permissions\": [\r\n        {\"perm\": \"permission\"},\r\n
    \       {\"perm\": \"permission.user\"},\r\n        {\"perm\": \"permission.org\"},\r\n
...
    \"操作成功\"\r\n}\r\n```"
...
  version: v0.0.1
paths:
...
  /mcp/server:
    delete:
      consumes:
      - application/json
      description: 删除MCP Server
      parameters:
      - description: mcpServerId
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.MCPServerIDReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 删除MCP Server
      tags:
      - tool
    get:
      consumes:
      - application/json
      description: 获取MCP Server详情
      parameters:
      - description: mcpServerId
        in: query
        name: mcpServerId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.MCPServerDetail'
              type: object
      summary: 获取MCP Server详情
      tags:
      - tool
    post:
      consumes:
      - application/json
      description: 创建MCP Server
      parameters:
      - description: MCP Server信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.MCPServerCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 创建MCP Server
      tags:
      - tool
    put:
      consumes:
      - application/json
      description: 更新MCP Server
      parameters:
      - description: MCP Server信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.MCPServerUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 更新MCP Server
      tags:
      - tool
  /mcp/server/list:
    get:
      consumes:
      - application/json
      description: 获取MCP Server列表
      parameters:
      - description: mcp server名称
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.ListResult'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/response.MCPServerInfo'
                        type: array
                    type: object
              type: object
      summary: 获取MCP Server列表
      tags:
      - tool
  /mcp/server/tool:
    delete:
      consumes:
      - application/json
      description: 删除MCP Server工具
      parameters:
      - description: mcpServerToolId
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.MCPServerToolIDReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 删除MCP Server工具
      tags:
      - tool
    post:
      consumes:
      - application/json
      description: 创建MCP Server工具
      parameters:
      - description: MCP Server工具信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.MCPServerToolCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 创建MCP Server工具
      tags:
      - tool
    put:
      consumes:
      - application/json
      description: 更新MCP Server工具
      parameters:
      - description: MCP Server工具信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.MCPServerToolUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 更新MCP Server工具
      tags:
      - tool
  /mcp/server/tool/custom/select:
    get:
      consumes:
      - application/json
      description: 获取MCP Server自定义工具列表（用于下拉选择）
      parameters:
      - description: CustomTool名称
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.ListResult'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/response.MCPServerCustomToolSelect'
                        type: array
                    type: object
              type: object
      summary: 获取MCP Server自定义工具列表（用于下拉选择）
      tags:
      - tool
  /mcp/server/tool/openapi:
    post:
      consumes:
      - application/json
      description: 创建openapi工具
      parameters:
      - description: openapi工具信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.MCPServerOpenAPIToolCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 创建openapi工具
      tags:
      - tool
...
  /tool/action/detail:
    get:
      consumes:
      - application/json
      description: 获取工具详情
      parameters:
      - description: action名称
        in: query
        name: actionName
        required: true
        type: string
      - description: 工具id
        in: query
        name: toolId
        required: true
        type: string
      - description: 工具类型
        enum:
        - builtin
        - custom
        in: query
        name: toolType
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.ToolActionDetail'
              type: object
      summary: 获取工具详情
      tags:
      - tool
  /tool/action/list:
    get:
      consumes:
      - application/json
      description: 获取工具列表
      parameters:
      - description: 工具id
        in: query
        name: toolId
        required: true
        type: string
      - description: 工具类型
        enum:
        - builtin
        - custom
        in: query
        name: toolType
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.ToolActionList'
              type: object
      summary: 获取工具列表
      tags:
      - tool
  /tool/builtin:
    post:
      consumes:
      - application/json
      description: 修改内置工具
      parameters:
      - description: 内置工具信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.ToolSquareAPIKeyReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 修改内置工具
      tags:
      - tool
  /tool/custom:
    delete:
      consumes:
      - application/json
      description: 删除自定义工具
      parameters:
      - description: 自定义工具ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.CustomToolIDReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 删除自定义工具
      tags:
      - tool
    get:
      consumes:
      - application/json
      description: 获取自定义工具详情
      parameters:
      - description: customToolId
        in: query
        name: customToolId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.CustomToolDetail'
              type: object
      summary: 获取自定义工具详情
      tags:
      - tool
    post:
      consumes:
      - application/json
      description: 创建自定义工具
      parameters:
      - description: 自定义工具信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.CustomToolCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 创建自定义工具
      tags:
      - tool
    put:
      consumes:
      - application/json
      description: 修改自定义工具
      parameters:
      - description: 自定义工具信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.CustomToolUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 修改自定义工具
      tags:
      - tool
  /tool/custom/list:
    get:
      consumes:
      - application/json
      description: 获取自定义工具列表
      parameters:
      - description: CustomTool名称
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.ListResult'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/response.CustomToolInfo'
                        type: array
                    type: object
              type: object
      summary: 获取自定义工具列表
      tags:
      - tool
  /tool/custom/schema:
    post:
      consumes:
      - application/json
      description: 解析自定义工具的Schema转换为API相关数据
      parameters:
      - description: Schema格式数据
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.CustomToolSchemaReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.ListResult'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/response.CustomToolActionInfo'
                        type: array
                    type: object
              type: object
      summary: 获取可用API列表（根据Schema）
      tags:
      - tool
  /tool/custom/select:
    get:
      consumes:
      - application/json
      description: 获取自定义工具列表（用于下拉选择）
      parameters:
      - description: CustomTool名称
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.ListResult'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/response.CustomToolSelect'
                        type: array
                    type: object
              type: object
      summary: 获取自定义工具列表（用于下拉选择）
      tags:
      - tool
  /tool/select:
    get:
      consumes:
      - application/json
      description: 获取工具列表（用于下拉选择）
      parameters:
      - description: 工具名
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.ListResult'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/response.ToolSelect'
                        type: array
                    type: object
              type: object
      summary: 获取工具列表（用于下拉选择）
      tags:
      - tool
  /tool/square:
    get:
      consumes:
      - application/json
      description: 获取内置工具详情
      parameters:
      - description: toolSquareId
        in: query
        name: toolSquareId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.ToolSquareDetail'
              type: object
      summary: 获取内置工具详情
      tags:
      - tool.square
  /tool/square/list:
    get:
      consumes:
      - application/json
      description: 获取内置工具列表
      parameters:
      - description: tool名称
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.ListResult'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/response.ToolSquareInfo'
                        type: array
                    type: object
              type: object
      summary: 获取内置工具列表
      tags:
      - tool.square
...
securityDefinitions:
  JWT:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"